backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cloudinary==1.33.0
Django==4.2.3
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.16
